Enable testing of JSON config generation/execution and stylistic improvements
