+++ .github/workflows/pytest-jsonconfig.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a single version of Python
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: pytest others
+name: pytest jsonconfig
 
 on:
   push:
@@ -15,7 +15,7 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python 3.11
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v2
       with:
         python-version: "3.11"
     - name: Setup Graphviz
@@ -29,8 +29,7 @@ jobs:
     - name: Test with pytest
       working-directory: ./tests/
       run: |
-        sh -c 'pytest -m "not base and not buildingtest and not system_setups and not utsp and not jsonconfig"; ret=$?; [ $ret = 5 ] && exit 0 || [ $ret = 0 ] && exit 0 || exit 1'
-        #        
+        pytest -m "jsonconfig" --durations=0
 
     #    # stop the build if there are Python syntax errors or undefined names
