+++ .gitignore
@@ -177,7 +177,6 @@ cfg.json
 tests/cfg.json
 tests/json/results
 json/results
-hisim/json_generator.py
 hisim/~$components_information.xlsx
 flake8_calls.txt
 prospector_calls.txt
