+++ hisim/json_executor.py
@@ -1,6 +1,7 @@
 """ Helper module to set up components and connections in the simulator based on the scenario data from the JSON file. """
-#clean
+# clean
 import importlib
+import os
 from typing import Any, cast
 from hisim import log, utils, loadtypes as lt
 from hisim.components.loadprofilegenerator_utsp_connector import UtspLpgConnector
@@ -39,6 +40,7 @@ def import_from_string(full_classname: str):
     except AttributeError as e:
         raise ImportError(f"Module '{module_path}' has no class '{class_name}'") from e
 
+
 def setup_components_and_connections(scenario_data: dict[str, Any], my_sim: sim.Simulator, sim_params: sim.SimulationParameters) -> sim.Simulator:
     """Given the scenario data from the JSON file, set up the components and connections in the simulator."""
 
@@ -66,9 +68,14 @@ def setup_components_and_connections(scenario_data: dict[str, Any], my_sim: sim.
                 config_dict["travel_route_set"] = humps.pascalize(config_dict["travel_route_set"])
                 config_dict["transportation_device_set"] = humps.pascalize(config_dict["transportation_device_set"])
                 config_dict["charging_station_set"] = humps.pascalize(config_dict["charging_station_set"])
+            elif comp_def["component_full_classname"] == "hisim.components.controller_l1_generic_ev_charge.L1Controller":
+                config_dict["charging_station_set"] = humps.pascalize(config_dict["charging_station_set"])
             elif comp_def["component_full_classname"] == "hisim.components.weather.Weather":
                 if "<<utils.get_input_directory()>>" in config_dict["source_path"]:
-                    config_dict["source_path"] = config_dict["source_path"].replace("<<utils.get_input_directory()>>", utils.get_input_directory())
+                    path_str = config_dict["source_path"].replace("<<utils.get_input_directory()>>", utils.get_input_directory())
+                    # Now replace "\\" and "/" with the correct os separator
+                    config_dict["source_path"] = os.path.normpath(path_str.replace("\\\\", os.sep).replace("\\", os.sep).replace("/", os.sep))
+                    # log.information(f"Resolved weather source path to {config_dict['source_path']}.")
                 else:
                     # This warning was generated by Copilot
                     log.warning(f"Unexpected value for source_path in Weather config: {config_dict['source_path']}.")
