+++ pytest.ini
@@ -8,3 +8,4 @@ markers =
     system_setups: systematic testing of the system_setups (deselect with '-m "not system_setups"')
     mpc: systematic testing of the mpc controlled systems (deselect with '-m "not mpc"')
     utsp: systematic testing of lpg utsp connector component (deselect with '-m "not utsp"')
+    jsonconfig: systematic testing of the json config generation and execution (deselect with '-m "not jsonconfig"')
