+++ tests/test_building_theoretical_thermal_demand.py
@@ -80,6 +80,7 @@ def test_house_with_idealized_electric_heater_for_heating_test(
     # Build Simulator
     normalized_path = os.path.normpath(PATH)
     path_in_list = normalized_path.split(os.sep)
+    path_to_be_added = None
     if len(path_in_list) >= 1:
         path_to_be_added = os.path.join(os.getcwd(), *path_in_list[:-1])
