+++ tests/test_json_configs.py
@@ -1,0 +1,52 @@
+"""Test for json config generation and execution."""
+from pathlib import Path
+import pytest
+from hisim import utils
+from hisim.hisim_convert_to_json import main
+import hisim.simulator as sim
+from hisim.hisim_main import initialize_from_python, initialize_from_json
+
+
+# This system setup generation from directory was generated by ChatGPT
+SYSTEM_SETUP_DIR = Path("../system_setups")
+SYSTEM_SETUPS = [
+    (str(path),)
+    for path in sorted(SYSTEM_SETUP_DIR.glob("*.py"))
+    if path.name != "__init__.py"
+]
+
+
+@pytest.mark.jsonconfig
+@pytest.mark.parametrize(["system_setup"], SYSTEM_SETUPS)
+@utils.measure_execution_time
+def test_json(system_setup: str):
+    """Test generation and execution of JSON configs."""
+
+    try:
+        sim_py = initialize_from_python(path_to_module=system_setup)
+    except Exception:
+        pytest.skip(f"Python configuration {system_setup} fails. Skipping JSON test for this configuration.")
+    else:
+        main(path_to_module=system_setup)
+        sim_json = initialize_from_json(scenario=system_setup[:-3] + ".scenario.json",
+                                        simulation_parameters=system_setup[:-3] + ".simulation.json",
+                                        path_to_module=system_setup[:-3] + ".scenario.json",
+                                        delta=None)
+        sim_py.prepare_calculation()
+        sim_json.prepare_calculation()
+        sim_py.connect_all_components()
+        sim_json.connect_all_components()
+
+        num_components_py, num_outputs_py = get_num_of_components_and_outputs(sim_py)
+        num_components_json, num_outputs_json = get_num_of_components_and_outputs(sim_json)
+
+        assert num_components_py == num_components_json
+        assert num_outputs_py == num_outputs_json
+
+
+def get_num_of_components_and_outputs(my_sim: sim.Simulator) -> tuple[int, int]:
+    """Helper function to get the number of components and outputs in the current simulator."""
+
+    num_components = len(my_sim.wrapped_components)
+    num_outputs = len(my_sim.all_outputs)
+    return num_components, num_outputs
